Add Node.js configuration in WizardNewTypeScriptProjectCreationPage - Node.js configuration is now in the first WizardPage of AbstractNewProjectWizard. - Removed Node.js configuration from TypeScriptRuntimeAndNodejsWizardPage and renamed it to TypeScriptRuntimeWizardPage.
