+++ eclipse/ts.eclipse.ide.ui/src/ts/eclipse/ide/internal/ui/TypeScriptUIMessages.properties
@@ -6,7 +6,7 @@
 # http://www.eclipse.org/legal/epl-v10.html
 #
 # Contributors:
-#     Angelo Zerr <angelo.zerr@gmail.com> - Initial API and implementation 
+#     Angelo Zerr <angelo.zerr@gmail.com> - Initial API and implementation
 #     Lorenzo Dalla Vecchia <lorenzo.dallavecchia@webratio.com> - Added save actions
 ###############################################################################
 
@@ -31,6 +31,9 @@ TypeScriptHyperlink_text=TypeScript - Go to definition
 TypeScriptConsoleJob_name=TypeScript Console Job
 ConsoleTerminateAction_tooltipText=Stop tsserver of [{0}] project.
 
+# Hover
+TypeScriptHover_openDeclaration=TypeScript - open declaration
+
 # Preferences
 TypeScriptMainPropertyPage_enable_builder_checkbox_label=Enable TypeScript B&uilder?
 TypeScriptMainPropertyPage_enable_builder_checkbox_description=Builder is required to <a href=\"compile\">compile/build on save</a> and <a href=\"tslint\">validate with tslint</a>.
@@ -42,7 +45,7 @@ TypeScriptMainPreferencePage_useSalsa_WhenNoJSDTNature=When no JSDT nature
 TypeScriptMainPreferencePage_refactoring_title= Refactoring TypeScript code
 TypeScriptMainPreferencePage_refactoring_auto_save= &Save all modified resources automatically prior to refactoring
 TypeScriptMainPreferencePage_refactoring_lightweight=Rename in editor &without dialog
-	
+
 PropertyAndPreferencePage_useworkspacesettings_change=Configure Workspace Settings...
 PropertyAndPreferencePage_showprojectspecificsettings_label=Configure Project Specific Settings...
 PropertyAndPreferencePage_useprojectsettings_label=Enable pr&oject specific settings
@@ -53,7 +56,7 @@ TypeScriptBuilder_enable_Error_message=Cannot add TypeScript Builder
 TypeScriptBuilder_disable_Error_message=Cannot remove TypeScript Builder
 
 # ATA
-ATAConfigurationBlock_description=Automatic Type Acquisition (ATA). 
+ATAConfigurationBlock_description=Automatic Type Acquisition (ATA).
 ATAConfigurationBlock_disableATA_checkbox_label=Disable ATA?
 ATAConfigurationBlock_enableTelemetry_checkbox_label=Enable Telemetry?
 
@@ -67,7 +70,7 @@ NodejsConfigurationBlock_embeddedNode_required_error=The Embedded node.js is req
 NodejsConfigurationBlock_installedNode_required_error=The installed node.js path is required.
 NodejsConfigurationBlock_nodeFile_exists_error=The ''{0}'' doesn't exists.
 NodejsConfigurationBlock_nodeFile_invalid_error=The ''{0}'' is not a node file.
- 
+
 # TypeScript Runtime
 TypeScriptRuntimeConfigurationBlock_typescript_group_label=Consume TypeScript with
 TypeScriptRuntimeConfigurationBlock_embedded_checkbox_label=Embedded TypeScript
@@ -81,7 +84,7 @@ TypeScriptRuntimeConfigurationBlock_embeddedTypeScript_required_error=The Embedd
 TypeScriptRuntimeConfigurationBlock_installedTypeScript_required_error=The installed TypeScript runtime path is required.
 TypeScriptRuntimeConfigurationBlock_typeScriptFile_exists_error=The ''{0}'' doesn't exists.
 TypeScriptRuntimeConfigurationBlock_typeScriptFile_invalid_error=The ''{0}'' is not a TypeScript Runtime.
-TypeScriptRuntimeConfigurationBlock_installedTypeScript_emulatePlugins_warning=No need to emulate plugins since TypeScript 2.2.1 support it. 
+TypeScriptRuntimeConfigurationBlock_installedTypeScript_emulatePlugins_warning=No need to emulate plugins since TypeScript 2.2.1 support it.
 
 # Formatter
 FormatterConfigurationBlock_editorOptions_group_label=Editor options
@@ -98,7 +101,7 @@ FormatterConfigurationBlock_formatOptions_insertSpaceAfterFunctionKeywordForAnon
 FormatterConfigurationBlock_formatOptions_insertSpaceAfterOpeningAndBeforeClosingNonemptyParenthesis= Defines space handling after opening and before closing non empty parenthesis.
 FormatterConfigurationBlock_formatOptions_insertSpaceAfterOpeningAndBeforeClosingNonemptyBrackets=Defines space handling after opening and before closing non empty brackets.
 FormatterConfigurationBlock_formatOptions_placeOpenBraceOnNewLineForFunctions=Defines whether an open brace is put onto a new line for functions or not.
-FormatterConfigurationBlock_formatOptions_placeOpenBraceOnNewLineForControlBlocks=Defines whether an open brace is put onto a new line for control blocks or not. 
+FormatterConfigurationBlock_formatOptions_placeOpenBraceOnNewLineForControlBlocks=Defines whether an open brace is put onto a new line for control blocks or not.
 
 # Save actions
 SaveActionsPreferencePage_performTheSelectedActionsOnSave=Perform the selected actions on save
@@ -171,7 +174,7 @@ TSConfigWizardPage_noImplicitReturns=No implicit returns
 TSConfigWizardPage_strictNullChecks=Strict null checks
 
 TypeScriptRuntimeAndNodejsWizardPage_title=TypeScript Runtime / Node.js
-TypeScriptRuntimeAndNodejsWizardPage_description=Select the TypeScript runtime and the Node.js required by the TypeScript tsserver.  
+TypeScriptRuntimeAndNodejsWizardPage_description=Select the TypeScript runtime and the Node.js required by the TypeScript tsserver.
 TypeScriptRuntimeAndNodejsWizardPage_nodejs_group_label=Node.js
 TypeScriptRuntimeAndNodejsWizardPage_typescript_group_label=TypeScript Runtime
 TypeScriptRuntimeAndNodejsWizardPage_useEmbeddedTsRuntime_label=Use embedded
